be/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Environment variable management dependencies
cryptography==41.0.7
pyyaml==6.0.1
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: KMS integration
boto3==1.34.0

# Optional: Slack integration
slack-sdk==3.23.0

# Optional: Background jobs
apscheduler==3.10.4

# Optional: RBAC
casbin==1.43.0
python-jose[cryptography]==3.3.0